Add README, with a sketch of a roadmap of sorts
The README describes what I would like this plugin to do, as a kind of
"successful outcome". I've also included a list at the bottom of steps
to take to achieve this. All in all, the ideas and goals are not well
defined, but I wanted to make sure that I got it all down.